Create README.md
The Food Killer project is a remake of the famous game, which contains additions to the gameplay mechanics.
Game process.
3 types of food fly out randomly. For cutting each, 1 point is added to the total score.
If the player does not have time to cut the food, it falls down and is out of the player's line of sight, life is taken away.
Only 3 lives. For each edible object that is not cut, 1 life is taken away.
Food is divided into 3 types - large, medium and small. When cutting a large object, 2 medium ones appear in its place,
in turn, when cutting the middle one, 2 small objects appear in its place.
There are bombs in the game - they cannot be cut. There are 3 types of bombs - large, medium and small.
When cutting a large bomb, 1 life is taken away, when cutting a medium bomb, 2 lives are taken away, when cutting a small bomb, all lives are lost
and the game is over.
The game ends only if all 3 lives are lost, the number of fruits and vegetables is not limited,
therefore, the number of points depends only on the skill of the player.
---------------------------------------------------------------------------------------------------------------------------------------------------
Проект «Food Killer» является ремейком известной игры «Fruit Ninja», в котором присутствуют дополнения в механике игрового процесса.
Процесс игры.
В рандомном порядке вылетают 3 вида еды. За разрезание каждого прибавляется 1 очко к общему счету. Если игрок не успевает разрезать еду, она падает вниз  и оказывается вне зоны видимости игрока, отнимается жизнь. Всего 3 жизни. За каждый не разрезанный съедобный объект отнимается по 1 жизни. 
Еда делится на 3 вида - большого, среднего и маленького размера. При разрезании большого объекта, на его месте появляется 2 средних, в свою очередь при разрезании среднего, на его место появляется 2 маленьких объекта. 
В игре присутствуют бомбы - их разрезать нельзя. Существует 3 вида бомб - большая, средняя и маленькая. При разрезании большой отнимается 1 жизнь, при разрезании средней отнимается 2 жизни, при разрезании маленькой бомбы - теряются все жизни и игра прекращается.
Игра заканчивается только в случае потери всех 3 жизней, количество фруктов и овощей не ограничено, поэтому количество очков зависит только от ловкости игрока.